+++ CHANGELOG.md
@@ -5,6 +5,8 @@
     - Function `neighbour_count`: Given an array (up to three
       dimensions), returns another array with the same shape containing
       the counts of cells' neighbours whose value is zero.
+    - Function `get_boundaries`: Given an array, returns either the mask
+      where the boundary edges are or their counts if specified.
 - `test_data_tools` has tests for almost all classes and functions.
   Functions that include plots are skipped (cannot be tested). Only test
   missing is `models.Lasso`.